// src/lib/game/enum.ts
/**
 * Enumeration representing mathematical operation symbols and a blank placeholder.
 * @enum {string}
 */
export enum MathSymbol {
  Addition = '+',
  Subtraction = '-',
  Multiplication = '*',
  Division = '/',
  Equals = '=',
  Blank = '?',
}

/**
 * Enum representing difficulty levels for the game.
 * @enum {string}
 */
export enum Difficulty {
  Easy = 'easy',
  Medium = 'medium',
  Hard = 'hard',
}